chore(scheduling): satisfy guard parity for the embedding-coverage job (BI-ED117C82)
Three preflight guards fired. All three were handled on their merits.

Module Size Guard — `catalog.ts` crossed the 800-LOC ceiling at 807 because of
the new entry. The catalog is a hand-maintained registry that grows with every
scheduled job, so capping the list would be the wrong lever; the SHAPE it
conforms to is what does not belong beside the data. `JobCategory` and
`ScheduledJobCatalogEntry` move to `catalog-types.ts` (pure types, no imports,
no runtime) and are re-exported so existing importers are untouched.
catalog.ts 807 -> 788.

Repo Guard Loop / no-bare-working-write — the guard flags `status: "working"`
written without `markTaskRunWorking`, to keep `lastHeartbeatAt` atomic so the
stall watchdog cannot false-positive. It matches the literal textually, and
this write is `Workroom.status`, not `TaskRun.status`: a durable maintenance
room has no heartbeat and no stall watchdog, and "working" is a valid member of
WORK_CAPSULE_STATUSES. Added to the guard's ALLOWLIST with that reason, which
is the path the guard itself documents for exactly this case.

Design Grounding — evidence recorded below.

## Design grounding

- Existing specs/plans reviewed:
  - docs/superpowers/plans/2026-08-16-embedding-coverage-self-heal.md (this work)
  - docs/superpowers/plans/2026-08-15-decision-review-honesty-and-retrieval.md §4 — the boot-hook version this completes
- Current code substrate reviewed:
  - apps/web/lib/wiki/embedding-reconciliation.ts — the reconcile plus `providerUnavailable`, already present
  - apps/web/lib/operate/scheduled-jobs/catalog.ts + scheduling-map.ts — the cron registry and same-tick contention read-model
  - apps/web/lib/queue/functions/business-journey-watchdog.ts — the pure-job + cron/run-now wrapper pattern followed here
  - packages/db/prisma/schema.prisma — Workroom / WorkroomActivity, the visibility surface
- Source of truth: the operator ruling that no end user should ever repair this, and that Workroom is the collection point for outcome-specific activity a coworker manages.
- Decision: extend the existing reconcile and the existing scheduled-job substrate; no new contract.

Design-Grounding-Decision: extend existing spec; no contract change.
